refactor: context 기반으로 컴포넌트 분리
